+++ Chat.IService/Interface/IUserService.cs
@@ -36,7 +36,7 @@ namespace Chat.IService.Interface
     public class UserSearchResult
     {
         public UserDTO[] Users { get; set; }
-        public int TotalCount { get; set; }
+        public long TotalCount { get; set; }
         public int WinCount { get; set; }
     }
 }
